Move the logic to add the LaaTransactionId defined in the header Laa-Transaction-Id to the AddLaaTransactionIdToMDC filter
