+++ maat-court-data-api/src/main/java/gov/uk/courtdata/assessment/controller/PassportAssessmentController.java
@@ -7,7 +7,6 @@ import gov.uk.courtdata.assessment.validator.PassportAssessmentValidationProcess
 import gov.uk.courtdata.dto.AssessorDetails;
 import gov.uk.courtdata.dto.PassportAssessmentDTO;
 import gov.uk.courtdata.eform.controller.StandardApiResponseCodes;
-import gov.uk.courtdata.enums.LoggingData;
 import gov.uk.courtdata.model.assessment.CreatePassportAssessment;
 import gov.uk.courtdata.model.assessment.UpdatePassportAssessment;
 import io.swagger.v3.oas.annotations.Operation;
@@ -50,7 +49,6 @@ public class PassportAssessmentController {
     public ResponseEntity<PassportAssessmentDTO> getAssessment(@PathVariable int passportAssessmentId,
                                                                @Parameter(description = "Used for tracing calls")
                                                                @RequestHeader(value = "Laa-Transaction-Id", required = false) String laaTransactionId) {
-        LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.debug("Get Passport Assessment Request Received");
         passportAssessmentValidationProcessor.validate(passportAssessmentId);
         PassportAssessmentDTO passportAssessment = passportAssessmentService.find(passportAssessmentId);
@@ -65,7 +63,6 @@ public class PassportAssessmentController {
     public ResponseEntity<PassportAssessmentDTO> getAssessmentByRepId(@PathVariable int repId,
                                                                       @Parameter(description = "Used for tracing calls")
                                                                       @RequestHeader(value = "Laa-Transaction-Id", required = false) String laaTransactionId) {
-        LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info("Get Passport Assessment by repId = {}", repId);
         PassportAssessmentDTO passportAssessment = passportAssessmentService.findByRepId(repId);
         return ResponseEntity.ok(passportAssessment);
@@ -80,7 +77,6 @@ public class PassportAssessmentController {
                     schema = @Schema(implementation = UpdatePassportAssessment.class))) @RequestBody UpdatePassportAssessment passportAssessment,
             @Parameter(description = "Used for tracing calls")
             @RequestHeader(value = "Laa-Transaction-Id", required = false) String laaTransactionId) {
-        LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.debug("Update Passport Assessment Request Received");
         passportAssessmentValidationProcessor.validate(passportAssessment);
         PassportAssessmentDTO updatedAssessment = passportAssessmentService.update(passportAssessment);
@@ -95,7 +91,6 @@ public class PassportAssessmentController {
             schema = @Schema(implementation = CreatePassportAssessment.class))) @RequestBody CreatePassportAssessment passportAssessment,
                                                                   @Parameter(description = "Used for tracing calls")
                                                                   @RequestHeader(value = "Laa-Transaction-Id", required = false) String laaTransactionId) {
-        LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.debug("Create Passport Assessment Request Received");
         passportAssessmentValidationProcessor.validate(passportAssessment);
         PassportAssessmentDTO newAssessment = passportAssessmentService.create(passportAssessment);
