+++ maat-court-data-api/src/main/java/gov/uk/courtdata/controller/ResultController.java
@@ -3,7 +3,6 @@ package gov.uk.courtdata.controller;
 import gov.uk.courtdata.annotation.NotFoundApiResponse;
 import gov.uk.courtdata.constants.CourtDataConstants;
 import gov.uk.courtdata.dto.ErrorDTO;
-import gov.uk.courtdata.enums.LoggingData;
 import gov.uk.courtdata.service.ResultsService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -42,7 +41,6 @@ public class ResultController {
             @PathVariable String asnSeq,
             @Parameter(description = "Used for tracing calls")
             @RequestHeader(value = CourtDataConstants.LAA_TRANSACTION_ID, required = false) String laaTransactionId) {
-        LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info(String.format("Get Result Codes by CaseId - %d and AsnSeq: %s {}", caseId, asnSeq));
         return ResponseEntity.ok(resultService.findResultCodesByCaseIdAndAsnSeq(caseId, asnSeq));
     }
