+++ maat-court-data-api/src/main/java/gov/uk/courtdata/controller/XLATResultController.java
@@ -3,7 +3,6 @@ package gov.uk.courtdata.controller;
 import gov.uk.courtdata.annotation.NotFoundApiResponse;
 import gov.uk.courtdata.constants.CourtDataConstants;
 import gov.uk.courtdata.dto.ErrorDTO;
-import gov.uk.courtdata.enums.LoggingData;
 import gov.uk.courtdata.service.ResultsService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -40,7 +39,6 @@ public class XLATResultController {
     public ResponseEntity<List<Integer>> getResultCodesForCCImprisonment(
             @Parameter(description = "Used for tracing calls")
             @RequestHeader(value = CourtDataConstants.LAA_TRANSACTION_ID, required = false) String laaTransactionId) {
-        LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info("Get XLAT Result Codes for CC Imprisonment");
         return ResponseEntity.ok(resultService.findXLATResultCodesForCCImprisonment());
     }
@@ -54,7 +52,6 @@ public class XLATResultController {
     public ResponseEntity<List<Integer>> getResultCodesForCCBenchWarrant(
             @Parameter(description = "Used for tracing calls")
             @RequestHeader(value = CourtDataConstants.LAA_TRANSACTION_ID, required = false) String laaTransactionId) {
-        LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info("Get XLAT Result Codes for CC Bench Warrant");
         return ResponseEntity.ok(resultService.findXLATResultCodesForCCBenchWarrant());
     }
@@ -70,7 +67,6 @@ public class XLATResultController {
             @PathVariable int subType,
             @Parameter(description = "Used for tracing calls")
             @RequestHeader(value = CourtDataConstants.LAA_TRANSACTION_ID, required = false) String laaTransactionId) {
-        LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info(String.format("Get XLAT Result Codes by WqType - %d and SubType: %s {}", wqType, subType));
         return ResponseEntity.ok(resultService.findXLATResultCodesByWQTypeAndSubTypeCode(wqType, subType));
     }
