+++ maat-court-data-api/src/main/java/gov/uk/courtdata/filter/AfterResponseClearDown.java
@@ -1,4 +1,4 @@
-package gov.uk.courtdata.controller;
+package gov.uk.courtdata.filter;
 
 import jakarta.servlet.Filter;
 import jakarta.servlet.FilterChain;
