+++ maat-court-data-api/src/main/java/gov/uk/courtdata/iojappeal/controller/IOJAppealController.java
@@ -3,7 +3,6 @@ package gov.uk.courtdata.iojappeal.controller;
 import gov.uk.courtdata.annotation.NotFoundApiResponse;
 import gov.uk.courtdata.dto.ErrorDTO;
 import gov.uk.courtdata.dto.IOJAppealDTO;
-import gov.uk.courtdata.enums.LoggingData;
 import gov.uk.courtdata.iojappeal.service.IOJAppealService;
 import gov.uk.courtdata.iojappeal.validator.IOJAppealValidationProcessor;
 import gov.uk.courtdata.model.iojAppeal.CreateIOJAppeal;
@@ -47,7 +46,6 @@ public class IOJAppealController {
     public ResponseEntity<IOJAppealDTO> getIOJAppeal(@PathVariable Integer iojAppealId,
                                                      @Parameter(description = "Used for tracing calls")
                                                      @RequestHeader(value = "Laa-Transaction-Id", required = false) String laaTransactionId) {
-      LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info("Get IOJ Appeal Received: id: {}", iojAppealId);
         return ResponseEntity.ok(iojAppealService.find(iojAppealId));
     }
@@ -61,7 +59,6 @@ public class IOJAppealController {
     public ResponseEntity<IOJAppealDTO> getIOJAppealByRepId(@PathVariable int repId,
                                                             @Parameter(description = "Used for tracing calls")
                                                             @RequestHeader(value = "Laa-Transaction-Id", required = false) String laaTransactionId) {
-      LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info("Get IOJ Appeal by repId: {}", repId);
         return ResponseEntity.ok(iojAppealService.findByRepId(repId));
     }
@@ -75,7 +72,6 @@ public class IOJAppealController {
     public ResponseEntity<IOJAppealDTO> getCurrentPassedIOJAppealByRepId(@PathVariable int repId,
                                                                          @Parameter(description = "Used for tracing calls")
                                                                          @RequestHeader(value = "Laa-Transaction-Id", required = false) String laaTransactionId) {
-      LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info("Get IOJ Appeal by repId: {}", repId);
         return ResponseEntity.ok(iojAppealService.findCurrentPassedAppealByRepId(repId));
     }
@@ -89,7 +85,6 @@ public class IOJAppealController {
             schema = @Schema(implementation = CreateIOJAppeal.class))) @Valid @RequestBody CreateIOJAppeal iojAppeal,
                                                         @Parameter(description = "Used for tracing calls")
                                                         @RequestHeader(value = "Laa-Transaction-Id", required = false) String laaTransactionId) {
-      LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info("Create IOJ Appeal Request Received");
 
         var iojAppealDTO = iojAppealService.create(iojAppeal);
@@ -106,7 +101,6 @@ public class IOJAppealController {
             schema = @Schema(implementation = UpdateIOJAppeal.class))) @Valid @RequestBody UpdateIOJAppeal iojAppeal,
                                                         @Parameter(description = "Used for tracing calls")
                                                         @RequestHeader(value = "Laa-Transaction-Id", required = false) String laaTransactionId) {
-      LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info("Update IOJ Appeal Request Received with ID: {}", iojAppeal.getId());
 
         iojAppealValidationProcessor.validate(iojAppeal);
