+++ maat-court-data-api/src/main/java/gov/uk/courtdata/reporder/controller/ReservationsController.java
@@ -4,7 +4,6 @@ import gov.uk.courtdata.annotation.NotFoundApiResponse;
 import gov.uk.courtdata.annotation.StandardApiResponse;
 import gov.uk.courtdata.constants.CourtDataConstants;
 import gov.uk.courtdata.entity.ReservationsEntity;
-import gov.uk.courtdata.enums.LoggingData;
 import gov.uk.courtdata.model.authorization.AuthorizationResponse;
 import gov.uk.courtdata.prosecutionconcluded.helper.ReservationsRepositoryHelper;
 import gov.uk.courtdata.reporder.service.ReservationsService;
@@ -49,7 +48,6 @@ public class ReservationsController {
             @PathVariable int maatId,
             @Parameter(description = "Used for tracing calls")
             @RequestHeader(value = CourtDataConstants.LAA_TRANSACTION_ID, required = false) String laaTransactionId) {
-        LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info(String.format("Check if maatId is locked - %d {}", maatId));
         return ResponseEntity.ok(reservationsRepositoryHelper.isMaatRecordLocked(maatId));
     }
@@ -62,7 +60,6 @@ public class ReservationsController {
             @PathVariable String recordName, @PathVariable Integer recordId,
             @Parameter(description = "Used for tracing calls")
             @RequestHeader(value = Constants.LAA_TRANSACTION_ID, required = false) String laaTransactionId) {
-        LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info("Check reservation status - request received");
         Optional<ReservationsEntity> reservationsEntity = reservationsRepositoryHelper.getReservationByRecordNameAndRecordId(recordName,recordId);
         return reservationsEntity.map(c -> ResponseEntity.ok().body(c))
@@ -76,7 +73,6 @@ public class ReservationsController {
     public ResponseEntity<ReservationsEntity> getReservationByUsername(
             @PathVariable String username,
             @RequestHeader(value = Constants.LAA_TRANSACTION_ID, required = false) String laaTransactionId) {
-        LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info("Check reservation status - request received");
         Optional<ReservationsEntity> reservationsEntity = reservationsRepositoryHelper.getReservationByUserName(username);
         return reservationsEntity.map(c -> ResponseEntity.ok().body(c))
