+++ maat-court-data-api/src/main/java/gov/uk/courtdata/unlink/controller/UnLinkController.java
@@ -1,7 +1,6 @@
 package gov.uk.courtdata.unlink.controller;
 
 import gov.uk.courtdata.dto.ErrorDTO;
-import gov.uk.courtdata.enums.LoggingData;
 import gov.uk.courtdata.model.Unlink;
 import gov.uk.courtdata.unlink.validator.UnLinkValidationProcessor;
 import io.swagger.v3.oas.annotations.Operation;
@@ -41,7 +40,6 @@ public class UnLinkController {
             @Parameter(description = "Case details data", content = @Content(mediaType = "application/json",
                     schema = @Schema(implementation = Unlink.class))) @RequestBody Unlink unlink) {
 
-        LoggingData.LAA_TRANSACTION_ID.putInMDC(laaTransactionId);
         log.info("LAA Status Update Request received");
         unLinkValidationProcessor.validate(unlink);
